Create Timeline/RaymarchingWorldSandwich
