+++ .idea/.idea.unity-demoscene/.idea/contentModel.xml
@@ -99,6 +99,14 @@
             <e p="Components" t="Include">
               <e p="RaymarchingWorldSandwichComponet.cs" t="Include" />
             </e>
+            <e p="Timeline" t="Include">
+              <e p="RaymarchingWolrdSandwich" t="Include">
+                <e p="RaymarchingWolrdSandwichBehaviour.cs" t="Include" />
+                <e p="RaymarchingWolrdSandwichClip.cs" t="Include" />
+                <e p="RaymarchingWolrdSandwichMixerBehaviour.cs" t="Include" />
+                <e p="RaymarchingWolrdSandwichTrack.cs" t="Include" />
+              </e>
+            </e>
           </e>
         </e>
         <e p="EffectExamples" t="Include">
