--- Assets/Demoscene/Scripts/Playables/RaymarchingWorldSandwich/RaymarchingWorldSandwichBehaviour.cs
@@ -1,16 +1,0 @@
-using System;
-using UnityEngine;
-using UnityEngine.Playables;
-using UnityEngine.Timeline;
-
-[Serializable]
-public class RaymarchingWorldSandwichBehaviour : PlayableBehaviour
-{
-    public GameObject newExposedReference;
-    public float newBehaviourVariable;
-
-    public override void OnGraphStart (Playable playable)
-    {
-        
-    }
-}
